Remove Heading circular dependency
Fixes #795
